Move property related folders under sidebar
