+++ source/frontend/src/features/mapSideBar/property/ComposedProperty.ts
