+++ source/frontend/src/features/mapSideBar/property/MotiInventoryContainer.test.tsx
