+++ source/frontend/src/features/mapSideBar/property/MotiInventoryContainer.tsx
@@ -1,6 +1,6 @@
 import { ReactComponent as LotSvg } from 'assets/images/icon-lot.svg';
 import GenericModal from 'components/common/GenericModal';
-import PropertyViewSelector from 'features/properties/map/propertyInformation/PropertyViewSelector';
+import PropertyViewSelector from 'features/mapSideBar/property/PropertyViewSelector';
 import SidebarFooter from 'features/properties/map/shared/SidebarFooter';
 import { FormikProps } from 'formik';
 import { PROPERTY_TYPES, useComposedProperties } from 'hooks/useComposedProperties';
@@ -8,7 +8,7 @@ import { Api_Property } from 'models/api/Property';
 import React, { useEffect, useRef, useState } from 'react';
 import styled from 'styled-components';
 
-import MapSideBarLayout from './layout/MapSideBarLayout';
+import MapSideBarLayout from '../layout/MapSideBarLayout';
 import { MotiInventoryHeader } from './MotiInventoryHeader';
 
 export interface IMotiInventoryContainerProps {
