+++ source/frontend/src/features/mapSideBar/property/MotiInventoryHeader.test.tsx
