+++ source/frontend/src/features/mapSideBar/property/PropertyContainer.tsx
