+++ source/frontend/src/features/mapSideBar/property/PropertyViewSelector.tsx
